+++ pandas/tests/series/methods/test_diff.py
@@ -1,0 +1,3 @@
+class TestDiff:
+    # TODO: maybe we should have tests for this?
+    pass
